# Add message and field name to binary serialization errors