This adds the message and field name to errors thrown during binary serialization of scalar fields. It should convert a message like:

```
Error: invalid uint32: -1
```

To:

```
Error: cannot encode field com.whatever.MyMessage.my_field to binary: invalid uint32: -1
```

I followed the wording of the error thrown here for the phrasing: https://github.com/bufbuild/protobuf-es/blob/main/packages/protobuf/src/to-binary.ts#L74

Fixes #983