@timostamm 2024-10-07
Thanks for the PR, Ben. We'll need some insight whether / how this impacts performance and bundle size. I think it would be ideal to use the same mechanism for error messages that `toJson` uses (but `writeField` is also used for extensions). We need good test coverage to make sure error messages have the expected content and will not regress.