@bhollis 2024-10-07
Running `npm run perf` in `packages/protobuf-test`:

Before:
```
fromBinary perf-payload.bin x 10,622 ops/sec ±0.70% (96 runs sampled)
fromBinary tiny example.User (4 bytes) x 1,243,008 ops/sec ±0.32% (99 runs sampled)
fromBinary normal example.User (92 bytes) x 407,844 ops/sec ±0.35% (97 runs sampled)
fromBinary scalar values (102 bytes) x 532,801 ops/sec ±2.21% (95 runs sampled)
fromBinary repeated scalar fields (195 bytes) x 341,618 ops/sec ±0.45% (96 runs sampled)
fromBinary map with scalar keys and values (186 bytes) x 202,318 ops/sec ±0.29% (94 runs sampled)
fromBinary repeated field with 1000 messages (2000 bytes) x 7,454 ops/sec ±1.98% (96 runs sampled)
fromBinary map field with 1000 messages (8890 bytes) x 2,063 ops/sec ±0.36% (96 runs sampled)
```

After:
```
fromBinary perf-payload.bin x 10,853 ops/sec ±0.44% (97 runs sampled)
fromBinary tiny example.User (4 bytes) x 1,225,170 ops/sec ±0.79% (97 runs sampled)
fromBinary normal example.User (92 bytes) x 379,662 ops/sec ±3.96% (91 runs sampled)
fromBinary scalar values (102 bytes) x 543,940 ops/sec ±1.85% (98 runs sampled)
fromBinary repeated scalar fields (195 bytes) x 332,102 ops/sec ±1.63% (94 runs sampled)
fromBinary map with scalar keys and values (186 bytes) x 193,957 ops/sec ±3.22% (95 runs sampled)
fromBinary repeated field with 1000 messages (2000 bytes) x 7,401 ops/sec ±3.26% (94 runs sampled)
fromBinary map field with 1000 messages (8890 bytes) x 2,052 ops/sec ±1.11% (98 runs sampled)
```