@bhollis 2024-12-18
A gentle nudge on this - it would still be of great benefit to have this information.